Suppress cast warnings.

+++ alvr_server/RecenterManager.h
@@ -111,7 +111,11 @@ public:
 
 		for (int i = 0; i < 2; i++) {
 			if (hapticFeedback[i][0] != 0 || hapticFeedback[i][1] != 0 || hapticFeedback[i][2] != 0) {
-				listener->SendHapticsFeedback(0, hapticFeedback[i][0], hapticFeedback[i][1], hapticFeedback[i][2], m_remoteController[i]->GetHand() ? 1 : 0);
+				listener->SendHapticsFeedback(0,
+					static_cast<float>(hapticFeedback[i][0]),
+					static_cast<float>(hapticFeedback[i][1]),
+					static_cast<float>(hapticFeedback[i][2]),
+					m_remoteController[i]->GetHand() ? 1 : 0);
 			}
 		}
